Update DX11 implem and fix with new VisualStudio

+++ src/Agmd3D/Core/Tools/Octree.h
@@ -142,7 +142,7 @@ namespace Agmd
 			std::stack<OcTree<T>*> stack;
 			if(!m_top)
 			{
-				float size = NearestPowerOfTwo(length(bsphere.GetRadius()+bsphere.GetCenter()));
+				float size = (float)NearestPowerOfTwo(length(bsphere.GetRadius()+bsphere.GetCenter()));
 				OcTree<T>* t = Allocate();
 				*t = OcTree<T>(this,BoundingBox(vec3(-size,-size,-size),vec3(size,size,size)));
 				m_top = t;
